feat: remove the content bip324/mod.rs into the existing modules
